added a aiming center and "invisible reticule" the dirrection and speed of the bullet is dictated by the dirrection of the vector between the center and reticule.
reticule is a child of the center, and when the player turns around the reticule turns 180 decreegs, turning again rotated the reticule -180 degrees. this is important as rotating more than 360 degrees has unexpected behavior.

NOTE TO SELF: add a safety, so that if a rotation puts the reticule over 359 degrees, it instead == 359